remove deprecated call which has been removed in SCons 4
I think this was an optimization from long ago, intended to prevent SCons
from taking time to check on whether the source tree was in sync with a
repo, but this call has been deprecated and finally removed in SCons 4.  It
was never actually needed in this tool, just an optimization.