Comment out install targets.
In theory: this should only be used as a submodule.  So all builds should be
in-tree.  People are starting to make Macports with aircraft_oap and the install
target is propgating down to these submodules in vardb.  Specifically there
was a comflict between xmlcatalog and libxml2 which had an xmlcatalog.